UI Consistency; Tidying; Setup Filenames
* Tidy wat/Translation.ini again (ensure *all* sections have two blank
lines preceding them, except the first one, which only has one, after
the header info).
* Tweak setup.exe filename to use underscores to separate components
and add "bit" to the end.
* Change the button for the about menu to just open the menu (the first
menu item already pointed to TPG's site, just as the button did, so it
was not necessary to add a menu item).
* Remove duplicate INI entries for the button for the About menu.
* Change the help link to go to the documentation.md on GitHub.
* Remove spaces around '='s in wat/Translation.ini.
* Remove duplicate INI entries for the IE_Options resid.
* Removed the button action to open the IE Settings window (the menu has
more features in and does not cause security warnings to pop up; see
the PR documentation for more info).